fastapi_aws_rds/requirements.txt
absl-py==0.11.0
affine==2.3.0
aiobotocore==1.1.2
aiohttp==3.7.3
aioitertools==0.7.1
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
Automat==20.2.0
awscli==1.18.201
awswrangler==2.0.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
bitarray==1.6.1
bleach==3.2.1
blis==0.7.3
boto3==1.16.39
botocore==1.19.41
bs4==0.0.1
cachetools==4.1.1
Cartopy==0.18.0
catalogue==1.0.0
cdsapi==0.3.1
certifi==2020.6.20
cffi==1.14.3
cfgrib==0.9.8.4
cftime==1.2.1
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.7.0
colorama==0.4.3
constantly==15.1.0
contextily==1.0.1
cryptography==3.2.1
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.4
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
descartes==1.1.0
docutils==0.15.2
en-core-web-sm==2.3.1
entrypoints==0.3
et-xmlfile==1.0.1
fastapi==0.63.0
fastparquet==0.4.1
filelock==3.0.12
Fiona==1.8.17
fsspec==0.8.4
future==0.18.2
fuzzywuzzy==0.18.0
gast==0.3.3
gcsfs==0.7.1
GDAL==3.0.0
GeoAlchemy2==0.8.4
geographiclib==1.50
geopandas==0.8.1
geopy==2.0.0
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.34.0
h11==0.11.0
h5py==2.10.0
html2json==0.2.4.1
httptools==0.1.1
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itemadapter==0.2.0
itemloaders==1.0.4
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
llvmlite==0.35.0
lxml==4.6.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
memory-profiler==0.58.0
mercantile==1.1.6
mistune==0.8.4
multidict==5.0.2
munch==2.5.0
murmurhash==1.0.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
netCDF4==1.5.4
nltk==3.5
notebook==6.1.5
numba==0.52.0
numpy==1.18.5
oauthlib==3.1.0
openpyxl==3.0.5
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
paramiko==2.7.2
parsel==1.6.0
parso==0.7.1
pdfreader==0.1.6
pexpect==4.8.0
pg8000==1.16.6
pickleshare==0.7.5
Pillow==8.0.1
plac==1.1.3
plotly==4.14.1
preshed==3.0.4
prometheus-client==0.8.0
prompt-toolkit==3.0.8
Protego==0.1.16
protobuf==3.14.0
psutil==5.7.3
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
pydantic==1.7.3
PyDispatcher==2.0.5
pyeccodes==0.1.1
Pygments==2.7.2
PyHamcrest==2.0.2
PyMySQL==0.10.1
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyproj==3.0.0.post1
pyquery==1.4.3
pyrsistent==0.17.3
pyshp==2.1.2
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
queuelib==1.5.0
rasterio==1.1.8
rasterstats==0.15.0
redshift-connector==2.0.872
regex==2020.11.13
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.5
s3fs==0.5.2
s3transfer==0.3.3
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.4
scramp==1.2.0
Scrapy==2.4.0
seaborn==0.11.0
Send2Trash==1.5.0
service-identity==18.1.0
Shapely==1.7.1
simplejson==3.17.2
six==1.15.0
sklearn==0.0
snuggs==1.4.7
soupsieve==2.0.1
spacy==2.3.4
SQLAlchemy==1.3.20
srsly==1.0.4
starlette==0.13.6
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
thinc==7.4.3
threadpoolctl==2.1.0
thrift==0.13.0
tokenizers==0.9.4
toml==0.10.2
torch==1.7.0
torchtext==0.8.0
tornado==6.1
tqdm==4.51.0
traitlets==5.0.5
transformers==4.0.0
Twisted==20.3.0
typing-extensions==3.7.4.3
urllib3==1.25.11
uvicorn==0.13.2
uvloop==0.14.0
VerbalExpressions==0.0.2
vulture==2.1
w3lib==1.22.0
wasabi==0.8.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
wget==3.2
widgetsnbextension==3.5.1
wrapt==1.12.1
xarray==0.16.1
xlrd==1.2.0
yarl==1.6.3
zope.interface==5.2.0
